Add files via upload
ITERR is a python module for solving the equations of motion for a charged particle moving in a 1D electric field which cam be modified to solve the same differential equations in 3D Electromagnetic field. Written in cpp to ensure the speed of obtaining solution and interfaced with python for the ease of use and visualizing the results using matplotlib.